% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi_banxico_data.R
\name{banxico_series2}
\alias{banxico_series2}
\title{Obtener datos de Banxico}
\usage{
banxico_series2(series, metadata = FALSE, verbose = FALSE, mask = FALSE)
}
\value{
regresa un vector de los montos deflactados
}
\description{
Esa función ayuda a extraer los datos de banxico. Está basada enteramente
en la librería de flores89/banxicoR, pero corrige un error que no permitía
descargar la información.
}
