% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_tools.R
\name{conect_value}
\alias{conect_value}
\title{Generar generar valor anterior para conectar líneas}
\usage{
conect_value(.x, variable, date = NA)
}
\arguments{
\item{.x}{Base de datos tidy}

\item{variable}{Variable en la que se basa el cambio}

\item{date}{Variable de fechas}
}
\value{
Regresa la base de datos con la colúmna con el índice
}
\description{
Esa función duplica el último valor de una serie integrando la categoría
de la siguiente serie
}
