% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\docType{data}
\name{CP2021}
\alias{CP2021}
\title{Datos del deflactor para la CP de 2021}
\format{
Data Frame
}
\source{
Datos del deflactor del PIB de la SHCP disponibles en los Lineamientos para la Integración de la Cuenta Pública 2021 (https://www.cuentapublica.hacienda.gob.mx/es/CP/cuenta)
}
\usage{
CP2021
}
\description{
Son los datos del deflactor implícito del PIB, observados y estimados
de la SHCP.
}
\keyword{datasets}
