% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_tools.R
\name{cve_to_ent}
\alias{cve_to_ent}
\title{Clave INEGI a entidad}
\usage{
cve_to_ent(cve_ent)
}
\arguments{
\item{cve_ent}{Clave de entidad}
}
\value{
un vector de caracteres de tamaño 3 de entidades de México no
ambiguas
}
\description{
Transforma las claves de INEGI correspondientes a cada entidad Mexicana en
nus nombres cortos. Si la clave no se encuentra regresa NA. Toda referencia a 
la Republica Federal o Nacion o Nacional se transformará en "Nacional" y
corresponderá a la clave 00..
}
