% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi_banxico_data.R
\name{deflactar_inpc}
\alias{deflactar_inpc}
\title{Deflactar montos INEGI}
\usage{
deflactar_inpc(monto, year_monto, year_out, mes = 12, acumulado = T)
}
\arguments{
\item{monto}{monto que se quiere deflactar}

\item{year_monto}{a침o de origen del monto}

\item{year_out}{a침o del precio al que se quiere deflactar}

\item{mes}{mes a comparar. En caso de ser acumulado sin que comience en enero,
tiene que ponerse una lista con el mes de inicio y fin.}

\item{acumulado}{variable l칩gica para determinar si se quiere acumulado}
}
\value{
regresa un vector de los montos deflactados
}
\description{
Esa funci칩n deflacta montos a partir del INPC.
}
