% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{deflactar_tp}
\alias{deflactar_tp}
\title{Deflactar montos}
\usage{
deflactar_tp(monto, year_monto, year_out, fuente = "local")
}
\arguments{
\item{monto}{monto que se quiere deflactar}

\item{year_monto}{año de origen del monto (para que funciones tiene que ser
entre 1994 y 2030)}

\item{year_out}{año del precio al que se quiere deflactar (para que funciones tiene que ser
entre 1994 y 2030)}

\item{fuente}{aquí se indica si se quiere tomar el deflactor que se guardó en el
paquete el 24 de junio de 2022 ("local") o se descarga el último archivo
 ("cargar"). TP lo actualiza posterior a la presentación de la estructura programática
 y el PPEF.}
}
\value{
regresa un vector de los montos deflactados
}
\description{
Esa función deflacta montos descargando la última versión de deflactores
publicada por Transparencia presupuestaria.
}
