% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\docType{data}
\name{deflactor_local}
\alias{deflactor_local}
\title{Datos del deflactor de TP}
\format{
Data Frame
}
\source{
Datos de TP disponibles en https://www.transparenciapresupuestaria.gob.mx/work/models/PTP/Presupuesto/Programacion/Deflactores/Deflactores_PIB.xlsx
}
\usage{
deflactor_local
}
\description{
Son los datos del deflactor implícito del PIB, observados y estimados
de la SHCP.
}
\keyword{datasets}
