% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{desc_ramo_to_abr_ramo}
\alias{desc_ramo_to_abr_ramo}
\title{Convertir nombre del ramo a abreviatura del ramo}
\usage{
desc_ramo_to_abr_ramo(desc_ramo)
}
\arguments{
\item{desc_ramo}{string con el nombre del ramo}
}
\value{
un string con la abreviatura del ramo
}
\description{
Convertir nombre del ramo a abreviatura del ramo
}
