% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_tools.R
\name{gen_index}
\alias{gen_index}
\title{Generar índice combaso en cambio entre periodos}
\usage{
gen_index(.x, col_from, col_to = index, ..., n_base = 100)
}
\arguments{
\item{.x}{Base de datos tidy}

\item{col_from}{Columna que contiene la tasa periodica de cambio (1 = 100%)}

\item{col_to}{Colúmna que se va a generar con el índice}

\item{...}{Componentes para seleccionar la base 100. Tiene que ser una obs.}

\item{n_base}{Base del índice}
}
\value{
Regresa la base de datos con la colúmna con el índice
}
\description{
Esa función genera un índice a partir de cambios de un periodo (ejemplo,
de un mes en estadisticas mensuales o de un trimestre en estadisticas
trimestrales)
}
