% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{gen_subclas_eco}
\alias{gen_subclas_eco}
\title{Generar subclasificación económica}
\usage{
gen_subclas_eco(.x, ..., periodo_col = NA, keep_mensual = T)
}
\arguments{
\item{.x}{base de datos de presupuesto}
}
\value{
una base de datos de presupuesto con subclasificación económica
}
\description{
Generar subclasificación económica
}
