% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{gen_tipo_programable}
\alias{gen_tipo_programable}
\title{Generar col\\u00famna con clasificaci\\u00f3n de programable}
\usage{
gen_tipo_programable(.x)
}
\arguments{
\item{.x}{base de datos de presupuesto}
}
\value{
una base de datos de presupuesto con clasificaci\\u00f3n del gasto
programable y no programable
}
\description{
Generar col\\u00famna con clasificaci\\u00f3n de programable
}
