% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\name{geom_segment_point}
\alias{geom_segment_point}
\title{Geom para generar segmentos sobre cada valor.}
\usage{
geom_segment_point(
  point_var_x,
  point_var_y,
  color_point = NULL,
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  point_dist = 0.4,
  ...,
  arrow = NULL,
  arrow.fill = NULL,
  lineend = "butt",
  linejoin = "round",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{data}{input1}
}
\value{
Una nueva capa de segmentos para una gráfica.
}
\description{
Esta función genera segmentos para cada valor de x y así mostrar cierto nivel.
}
