% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\name{geom_text_bilateral}
\alias{geom_text_bilateral}
\title{Geom text para barras horizontales}
\usage{
geom_text_bilateral(
  lab_position,
  lab_hor = NULL,
  percent_change = 0.25,
  hdist = 0.05,
  vdist = NULL,
  vjust = 0.5,
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  color_black = "grey15",
  color_light = "grey95",
  ...,
  parse = FALSE,
  nudge_x = 0,
  nudge_y = 0,
  check_overlap = FALSE,
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{data}{input1}
}
\value{
Una nueva capa de texto para una gráfica
}
\description{
Esta función sirve para añadir fácilmente los valores para los barras de variación
en valores negativos y positivos. Sólo se esplican los parámetros que difieren de
geom_text() normal
}
