% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_objects.R
\name{guardar_obj}
\alias{guardar_obj}
\title{Guardar base de datos en nuevo objeto}
\usage{
guardar_obj(data, nombre = "obj_guardado")
}
\arguments{
\item{data}{cualquier objeto}

\item{nombre}{string que se le asigna como nombre al objeto. Si no se
asigna ninguno, se guardará como obj_guardado .}
}
\value{
el mismo objeto y guarda un nuevo objeto con el nombre de bd_guardada
}
\description{
Esta función sirve para guardar un objeto  tal como está en cierto punto
de una secuencia de pipas.
}
