% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{id_capitulo_to_desc_capitulo}
\alias{id_capitulo_to_desc_capitulo}
\title{ID de capítulo a descripción de capítulo}
\usage{
id_capitulo_to_desc_capitulo(id_ramo)
}
\arguments{
\item{id_capitulo}{la colúmna con la id del capítulo.}
}
\value{
regresa una lista con la descripción del capitulo
}
\description{
Esta función regresa la descripción del capítulo poniendo el capítulo correspondiente
}
