% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{id_concepto_to_desc_concepto}
\alias{id_concepto_to_desc_concepto}
\title{ID de concepto a descripción de concepto}
\usage{
id_concepto_to_desc_concepto(id_concepto)
}
\arguments{
\item{id_concepto}{la colúmna con la id del concepto.}
}
\value{
regresa una lista con la descripción del concepto.
}
\description{
Esta función regresa la descripción del concepto poniendo el concepto correspondiente
}
