% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{id_part_esp_to_desc_part_esp}
\alias{id_part_esp_to_desc_part_esp}
\title{ID de partida específica a descripción de partida específica}
\usage{
id_part_esp_to_desc_part_esp(id_partida_especifica)
}
\arguments{
\item{id_partida_especifica}{la colúmna con la id del partida específica.}
}
\value{
regresa una lista con la descripción del partida específica.
}
\description{
Esta función regresa la descripción del partida específica poniendo el partida específica correspondiente
}
