% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{id_part_gen_to_desc_part_gen}
\alias{id_part_gen_to_desc_part_gen}
\title{ID de partida genérica a descripción de partida genérica}
\usage{
id_part_gen_to_desc_part_gen(id_partida_generica)
}
\arguments{
\item{id_partida_generica}{la colúmna con la id del partida genérica.}
}
\value{
regresa una lista con la descripción del partida genérica.
}
\description{
Esta función regresa la descripción del partida genérica poniendo el partida genérica correspondiente
}
