% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{id_ramo_to_abr_ramo}
\alias{id_ramo_to_abr_ramo}
\title{Convertir clave del ramo a abreviatura del ramo}
\usage{
id_ramo_to_abr_ramo(id_ramo)
}
\arguments{
\item{id_ramo}{clave númerica del ramo}
}
\value{
character con la clasificación administrativa del ramo
}
\description{
Nota 1: esta clasificación a la última actualización del año fiscal.
Por ejemplo, en esta clasificación ya se considera al ramo 49 (FGR)
como Fiscalía y no Procuraduría.
}
