% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{negative_neteo_tp}
\alias{negative_neteo_tp}
\title{Presupuesto para neteo en negativo}
\usage{
negative_neteo_tp(.x)
}
\arguments{
\item{.x}{.xframe con la estructura de los datos abiertos de
presupuesto de Transparencia Presupuestaria.}
}
\value{
.xframe cuyas categor\\u00edas del gasto con descripci\\u00f3n del ramo
como "Neteo" est\\u00e1n en valores negativos.
}
\description{
Esta funci\\u00f3n convierte aquellas categor\\u00edas del presupuesto que tienen
"Neteo" como descripci\\u00f3n del ramo en valores negativos. Normalmente se
deber\\u00eda usar despu\\u00e9s de la funci\\u00f3n de netear_tp.
}
