% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presupuestoR-package.R
\docType{package}
\name{presupuestoR-package}
\alias{presupuestoR}
\alias{presupuestoR-package}
\title{presupuestoR: Paquete para análisis del presupuesto federal de México}
\description{
Este paquete contiene funciones para procesar de manera más eficiente la información presupuestalmente de México. Particularmente en el formato de Transparencia Presupuestaria.
}
\keyword{internal}
