% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{sum_pef_tp}
\alias{sum_pef_tp}
\title{Agrupaci\\u00f3n del presupuesto}
\usage{
sum_pef_tp(.x, ..., periodo_col = NA, keep_mensual = T)
}
\arguments{
\item{.x}{.x frame con la estructura de los archivos sobre presupuesto
de transparencia Presupuestaria
(https://www.transparenciapresupuestaria.gob.mx/es/PTP/Datos_Abiertos).}

\item{...}{nombre de las variables por las cuales se quiere colapsar
la informaci\\u00f3n. Estos nombre tienen que estar completamente en min\\u00fascula
sin caracteres especiales y separado por guiones bajos. Por ejemplo,
id_ramo, desc_ramo, id_modalidad, desc_modalidad, etc.}

\item{periodo_col}{agregar columna con el periodo correspondiente}

\item{keep_mensual}{indicador lógico para determinar si mantener la variables de aprobado
y modificado en el periodo (en caso de estar diponible)}
}
\value{
.xframe colapsado por las variables indicadas. En caso de no
indicarse ninguna variable, s\\u00f3lo se colapsar\\u00e1 por a\\u00f1o fiscal.
}
\description{
Esta funci\\u00f3n permite agrupar las bases de datos abiertas del presupuesto
de Transparencia Presupuestaria con los principales niveles de agregaci\\u00f3n.
De esta manera, podemos generar resumenes presupuestales de acuerdo a su
clasificaci\\u00f3n administrativa, funcional o econ\\u00f3mica.
Por ejemplo, permite generar r\\u00e1pidamente un .xframe con el presupuesto
en todas sus categor\\u00edas disponibles(aprobado, modificado, pagado y
ejercido) por ramo.
Nota 1: esta funci\\u00f3n siempre mantiene el periodo indicado y, en caso de no
tenerlo, toma el a\\u00f1o fiscal como periodo.
Nota 2: esta funci\\u00f3n no distingue los cambios de claves y t\\u00e9rminos a lo
largo del tiempo. Por ejemplo, si un programa cambi\\u00f3 de modalidad U a S,
e intentamos agrupar el programa con su modalidad y clave presupuestal,
la funci\\u00f3n la va distinguir como un programa diferente. Otro ejemplo es que
esta funci\\u00f3n no distingue  los cambios en el tiempo de la en la
clasificaci\\u00f3n por objeto del gasto. Si una misma categor\\u00eda tiene una
clave o un nombre distinto a lo largo del tiempo y se clasifica el
presupuesto por estas categor\\u00edas, la funci\\u00f3n lo clasificar\\u00e1 como
categor\\u00edas diferentes.
}
