% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_objects.R
\name{unrowwise_df}
\alias{unrowwise_df}
\title{Desagrupar por fila}
\usage{
unrowwise_df(data)
}
\arguments{
\item{data}{dataframe agrupado por fila}
}
\value{
dataframe sin agrupación
}
\description{
Esta función sirve para desagrupar un dataframe agrupado por
fila (dplyr::rowwise()).
}
