% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_geoms.R
\name{width_bar}
\alias{width_bar}
\title{Valores absolutos abarcados}
\usage{
width_bar(variable)
}
\arguments{
\item{variable}{Vector o colúmna de datos a partir del cuál se calculan los datoss}
}
\value{
Valor absoluto de los valores absolutos (positivos y negativos) que recorre
el vector o colúmna.
}
\description{
Esta función te muestra en valores absolutos los montos recorridos por la colúmna
o vector evaluados.
}
