% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi_banxico_data.R
\name{banxico_series2}
\alias{banxico_series2}
\title{Obtener datos de Banxico}
\usage{
banxico_series2(series, metadata = FALSE, verbose = FALSE, mask = FALSE)
}
\arguments{
\item{series}{clave de la serio}

\item{metadata}{indicador lógico sobre la inclusión metadara}

\item{verbose}{indicador lógico para señalar si muestra los pasos}

\item{mask}{comando lógico si usar el nombre de la serie o values como columna}
}
\value{
regresa un vector de los montos deflactados
}
\description{
Esa función ayuda a extraer los datos de banxico. Está basada enteramente
en la librería de flores89/banxicoR, pero corrige un error que no permitía
descargar la información.
}
