% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{bind_pef_tp_wide}
\alias{bind_pef_tp_wide}
\title{Agrupar lista de dataframes de presupuesto en formato wide}
\usage{
bind_pef_tp_wide(lista_df, ...)
}
\arguments{
\item{lista_df}{lista de dataframes con la estructura de los datos abiertos
de presupuesto de Transparencia Presupuestaria.}

\item{...}{nombre de las variables por las cuales se quiere colapsar
la información. Estos nombre tienen que estar completamente en minúscula
sin caracteres especiales y separado por guiones bajos. Por ejemplo,
id_ramo, desc_ramo, id_modalidad, desc_modalidad, etc.}
}
\value{
dataframe colapsado por las variables indicadas. En caso de no
indicarse ninguna variable, sólo se colapsará por año fiscal.
}
\description{
Esta función retoma la una función sum_pef_tp (por lo cual le aplican las
mismas reglas) y la aplica a una lista de presupuesto. Esto ayuda cuando
se quiere generar datos históricos del presupuesto en formato wide, los
cuales normalmente están en diferentes archivos.
Lo ideal es cargar todos los dataframes de presupuesto en una lista de
dataframes y seleccionar las variables por las que se quiere acumular la
información.
Por ejemplo. Si colocamos un vector de las cuentas públicas de cada año
fiscal sin señalar ninguna variable, esta función regresará el total
del presupuesto (sin neteo) de todos los ciclos fiscales incluídos.
Nota 1: si se acumulan dos dataframes del mismo año fiscal y no de le
distingue con una variable periodo, esta función sólo señalará el ciclo
fiscal sin distinguir si son de trimestress diferentes.
Nota 2: esta función no distingue los cambios de claves y términos a lo
largo del tiempo. Por ejemplo, si un programa cambió de modalidad U a S,
e intentamos agrupar el programa con su modalidad y clave presupuestal,
la función la va distinguir como un programa diferente. Otro ejemplo es que
esta función no distingue  los cambios en el tiempo de la en la
clasificación por objeto del gasto. Si una misma categoría tiene una
clave o un nombre distinto a lo largo del tiempo y se clasifica el
presupuesto por estas categorías, la función lo clasificará como
categorías diferentes.
}
