% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inegi_banxico_data.R
\name{deflactar_inpc}
\alias{deflactar_inpc}
\title{Deflactar montos INEGI}
\usage{
deflactar_inpc(monto, year_monto, year_out, mes = 12)
}
\arguments{
\item{monto}{monto que se quiere deflactar}

\item{year_monto}{año de origen del monto}

\item{year_out}{año del precio al que se quiere deflactar}

\item{mes}{mes a comparar}
}
\value{
regresa un vector de los montos deflactados
}
\description{
Esa función deflacta montos a partir del INPC.
}
