% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{gen_clas_eco}
\alias{gen_clas_eco}
\title{Generar clasificación económica}
\usage{
gen_clas_eco(data, ..., periodo_col = NA, keep_mensual = T)
}
\arguments{
\item{.x}{base de datos de presupuesto}
}
\value{
una base de datos de presupuesto con clasificación económica
}
\description{
Generar clasificación económica
}
