% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{gen_subclas_eco}
\alias{gen_subclas_eco}
\title{Generar subclasificación económica}
\usage{
gen_subclas_eco(.x, name_clas = clasif_eco, name_sub = sub_clasif_eco)
}
\arguments{
\item{.x}{base de datos de presupuesto}

\item{name_clas}{nombre de la columna con la clasificación económica}

\item{name_sub}{nombre de la columna con la sclasificación económica}
}
\value{
una base de datos de presupuesto con subclasificación económica
}
\description{
Generar subclasificación económica
}
