% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{gen_subclas_eco}
\alias{gen_subclas_eco}
\title{Generar subclasificaci\\u00f3n econ\\u00f3mica}
\usage{
gen_subclas_eco(.x, name_clas = clasif_eco, name_sub = sub_clasif_eco)
}
\arguments{
\item{.x}{base de datos de presupuesto}

\item{name_clas}{nombre de la columna con la clasificaci贸n econ贸mica}

\item{name_sub}{nombre de la columna con la sclasificaci贸n econ贸mica}
}
\value{
una base de datos de presupuesto con subclasificaci\\u00f3n econ\\u00f3mica
}
\description{
Generar subclasificaci\\u00f3n econ\\u00f3mica
}
