% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{gen_tipo_programable}
\alias{gen_tipo_programable}
\title{Generar colúmna con clasificación de programable}
\usage{
gen_tipo_programable(.x)
}
\arguments{
\item{.x}{base de datos de presupuesto}
}
\value{
una base de datos de presupuesto con clasificación del gasto
programable y no programable
}
\description{
Generar colúmna con clasificación de programable
}
