% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{id_ramo_to_abr_ramo}
\alias{id_ramo_to_abr_ramo}
\title{Convertir clave del ramo a abreviatura del ramo}
\usage{
id_ramo_to_abr_ramo(id_ramo)
}
\arguments{
\item{id_ramo}{clave n\\u00famerica del ramo}
}
\value{
character con la clasificaci\\u00f3n administrativa del ramo
}
\description{
Nota 1: esta clasificaci\\u00f3n a la \\u00faltima actualizaci\\u00f3n del a\\u00f1o fiscal.
Por ejemplo, en esta clasificaci\\u00f3n ya se considera al ramo 49 (FGR)
como Fiscal\\u00eda y no Procuradur\\u00eda.
}
