% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistic_tools.R
\name{indexing}
\alias{indexing}
\title{Indexar o reindexar}
\usage{
indexing(.x, col_from, col_to = index, ..., n_base = 100)
}
\arguments{
\item{.x}{Base de datos tidy}

\item{col_from}{Columna que se va a indexar}

\item{col_to}{Colúmna que se va a generar con el resultado}

\item{...}{Componentes a filtrar}

\item{n_base}{Base del índice}
}
\value{
Regresa la base de datos con la colúmna con el índice
}
\description{
Esa función genera un índice a partir de cierto periodo base.
}
