% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{negative_neteo_tp}
\alias{negative_neteo_tp}
\title{Presupuesto para neteo en negativo}
\usage{
negative_neteo_tp(data)
}
\arguments{
\item{data}{dataframe con la estructura de los datos abiertos de
presupuesto de Transparencia Presupuestaria.}
}
\value{
dataframe cuyas categorías del gasto con descripción del ramo
como "Neteo" están en valores negativos.
}
\description{
Esta función convierte aquellas categorías del presupuesto que tienen
"Neteo" como descripción del ramo en valores negativos. Normalmente se
debería usar después de la función de netear_tp.
}
