% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{netear_tp}
\alias{netear_tp}
\title{Netear}
\usage{
netear_tp(data, ...)
}
\arguments{
\item{data}{dataframe con la estructura de los datos abiertos de
presupuesto de Transparencia Presupuestaria.}

\item{...}{modificación y creación de variables equivalente a dplyr::mutate}
}
\value{
dataframe cuya descripción de ramo indica los casos en que esa
categoría del gasto debe ser neteada.
}
\description{
Esta función pretende clasificar categoría del presupuesto que deben ser
neteadas en caso de realizarse un análisis del presupuesto neto. Esta
clasificación la realiza sobre archivos presupuestales con la estructura
de los datos de presupuesto abierto de Transparencia Presupuestaria.
Ojo: esta función aún está en desarrollo y, hasta ahora, el neteo sólo
es correcto aplicado al presupuesto aprobado.
}
