% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisis_ppto.R
\name{sum_pef_tp}
\alias{sum_pef_tp}
\title{Agrupación del presupuesto}
\usage{
sum_pef_tp(data, ...)
}
\arguments{
\item{data}{data frame con la estructura de los archivos sobre presupuesto
de transparencia Presupuestaria
(https://www.transparenciapresupuestaria.gob.mx/es/PTP/Datos_Abiertos).}

\item{...}{nombre de las variables por las cuales se quiere colapsar
la información. Estos nombre tienen que estar completamente en minúscula
sin caracteres especiales y separado por guiones bajos. Por ejemplo,
id_ramo, desc_ramo, id_modalidad, desc_modalidad, etc.}
}
\value{
dataframe colapsado por las variables indicadas. En caso de no
indicarse ninguna variable, sólo se colapsará por año fiscal.
}
\description{
Esta función permite agrupar las bases de datos abiertas del presupuesto
de Transparencia Presupuestaria con los principales niveles de agregación.
De esta manera, podemos generar resumenes presupuestales de acuerdo a su
clasificación administrativa, funcional o económica.
Por ejemplo, permite generar rápidamente un dataframe con el presupuesto
en todas sus categorías disponibles(aprobado, modificado, pagado y
ejercido) por ramo.
Nota 1: esta función siempre mantiene el periodo indicado y, en caso de no
tenerlo, toma el año fiscal como periodo.
Nota 2: esta función no distingue los cambios de claves y términos a lo
largo del tiempo. Por ejemplo, si un programa cambió de modalidad U a S,
e intentamos agrupar el programa con su modalidad y clave presupuestal,
la función la va distinguir como un programa diferente. Otro ejemplo es que
esta función no distingue  los cambios en el tiempo de la en la
clasificación por objeto del gasto. Si una misma categoría tiene una
clave o un nombre distinto a lo largo del tiempo y se clasifica el
presupuesto por estas categorías, la función lo clasificará como
categorías diferentes.
}
